modelo de calculo pruebas

+++ mono-01/src/mono_01/pipelines/data_mono/nodes.py
@@ -54,7 +54,7 @@ def imputar_datos(df_filtered: pd.DataFrame) -> pd.DataFrame:
     # Reemplazar las columnas imputadas en el DataFrame original filtrado
     df_filtered[columnas_para_imputar] = df_filtered_imputed
     # Convertir todas las columnas imputadas a int64
-    df_filtered[columnas_para_imputar] = df_filtered[columnas_para_imputar].fillna(0).astype('int64')
+    df_filtered[columnas_para_imputar] = df_filtered[columnas_para_imputar].fillna(0).astype('int32')
     return df_filtered
 
 def pre_proceso(df: pd.DataFrame) -> pd.DataFrame:
